PoseTrack/benchmarks/requirements_bench.txt
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
pillow==10.4.0
tensorflow==2.16.2
tensorflow-hub==0.16.1
